+++ vm.c
@@ -519,7 +519,7 @@ static void value_obj_free(Vm* vm) {
     }
 }
 
-Result vm_interpret(char* source, ssize_t source_len,
+Result vm_interpret(char* source, size_t source_len,
                     Result (*bytecode_fn)(Vm*, Chunk*)) {
     Vm vm = {.globals = ht_init(100, NULL)};
     Chunk chunk = {0};
@@ -566,8 +566,8 @@ void vm_repl(void) {
         Chunk chunk = {0};
         Result result = RES_OK;
 
-        if ((result = parser_compile(source, source_len, &chunk, &vm)) !=
-            RES_OK)
+        if ((result = parser_compile(source, (size_t)source_len, &chunk,
+                                     &vm)) != RES_OK)
             continue;
 
         LOG("parsing successful%s\n", "");
